+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.14.0+2 (not released yet)
+## 1.15.0 (2020-03-30)
 * API change: Changed `submittable` type from `bool` to `SubmittableMeta` in the `ProvidedMeta`.
 * Added an `activatable` flag to an action.
 * Added `features` to the request header.